Update logging mechanism to use log4j
For an example of how to log messages with log4j, see the Logic.java file in the server src package. Note that messages logged with this system appear on the common.Console window, as well as the standard eclipse console window. The messages are also immediately logged to a file eliminating the need to do it when the application shuts down. The settings for logging threshold, log file name and location, max size of log files etc, are all configurable from the .properties files in the Log Settings folder, and can be changed without needing to recompile the code. The jar file for log4j has been added under a Lib directory and must be referenced to compile the code.