added basic support for CommandRejected, need to implement an undo system.
added None to restriction category to be received when calling getRestriction and size is zero. None does not count in hasRestriction. this is added to prevent exceptions in switch cases.

removed firing of SpecialCharUpdate from setupPhase, to be only called when client can access special characters.

fixed some animations in the GUI.